reddit-alerts/requirements.txt
praw
arrow
marshmallow
marshmallow-sqlalchemy
